created PG Pool object as module in pool.js file
