+++ server/server.js
@@ -2,7 +2,7 @@ const express = require('express');
 const app = express();
 const bodyParser = require('body-parser');
 const PORT = process.env.PORT || 5000;
-const koalaRouter = require('./routes/koala.router');
+const koalaRouter = require('./routes/koala.router')
 
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(express.static('server/public'));
